feat: Phase 2 - migrate modules to package structure
Complete module migration from modules/ to src/index_pdf_translation/:

## Module Migration
- modules/logger.py → src/.../logger.py
- modules/spacy_api.py → src/.../nlp/tokenizer.py (renamed)
- modules/pdf_edit.py → src/.../core/pdf_edit.py
- modules/translate.py → src/.../core/translate.py

## New Files
- src/.../config.py: Language configuration (SUPPORTED_LANGUAGES)
- src/.../resources/fonts/: Font files (Liberation Serif, IPA Mincho)
- src/.../resources/data/: Logo and assets

## Import Updates
- All internal imports updated to new package paths
- translate_pdf.py updated to import from index_pdf_translation
- Package __init__.py now exports pdf_translate

## Deleted
- modules/ directory (migrated to src/)

## Verification
- `from index_pdf_translation import pdf_translate` works
- `from index_pdf_translation.nlp import tokenize_text` works
- CLI `translate_pdf.py --help` works

Relates to #17
