+++ src/index_pdf_translation/resources/data/__init__.py
@@ -1,0 +1,2 @@
+# SPDX-License-Identifier: AGPL-3.0-only
+"""Data resources (logo, etc.)."""
