+++ src/index_pdf_translation/resources/fonts/__init__.py
@@ -1,0 +1,2 @@
+# SPDX-License-Identifier: AGPL-3.0-only
+"""Font resources for PDF text rendering."""
